+++ src/Graphics/Geometry.hpp
@@ -11,7 +11,7 @@ struct Vertex {
   glm::vec3 position;
   glm::vec3 normal;
   glm::vec2 uv;
-  float materialId;
+  int32_t materialId;
 };
 
 class Geometry {
